Fix: Plugin overview page showed all versions as 0.1.2.
The nprogress jQuery version function was being called with the effect that
any element with the class "version" was being set to "0.1.2".
https://github.com/bmbrands/moodle-theme_elegance/issues/78

I can't see that this function is required to use nprogress, the progress bar
appears to work fine without it, so removed this call which fixed the issue for
me.